chore: remove unused files and move hooks and utils to correct dirs
