+++ libs/deal-ticket/src/components/deal-ticket-validation/index.ts
@@ -1,5 +1,1 @@
-export * from './get-default-order';
-export * from './use-order-validation';
-export * from './validate-amount';
-export * from './use-persisted-order';
-export * from './validate-expiration';
+export * from './validate-margin';
