+++ libs/deal-ticket/src/hooks/use-order-margin-validation.ts
@@ -2,9 +2,9 @@ import { useMemo } from 'react';
 import { useVegaWallet } from '@vegaprotocol/wallet';
 import { Schema } from '@vegaprotocol/types';
 import { toBigNum } from '@vegaprotocol/react-helpers';
-import type { DealTicketMarketFragment } from '../deal-ticket/__generated__/DealTicket';
-import type { OrderMargin } from '../../hooks/use-order-margin';
-import { usePartyBalanceQuery, useSettlementAccount } from '../../hooks';
+import type { DealTicketMarketFragment } from '../components/deal-ticket/__generated__/DealTicket';
+import type { OrderMargin } from './use-order-margin';
+import { usePartyBalanceQuery, useSettlementAccount } from '.';
 
 interface Props {
   market: DealTicketMarketFragment;
