+++ libs/deal-ticket/src/utils/validate-expiration.ts
