+++ libs/deal-ticket/src/utils/validate-market-state.ts
