+++ libs/deal-ticket/src/utils/validate-market-trading-mode.ts
